refactor: 코드 스타일 및 주석 삭제
