+++ tomcat/src/main/java/org/apache/catalina/startup/Server.java
@@ -1,4 +1,4 @@
-package com.techcourse;
+package org.apache.catalina.startup;
 
 public interface Server {
     void start();
