+++ tomcat/src/main/java/org/apache/catalina/startup/WAS.java
@@ -1,8 +1,9 @@
-package com.techcourse;
+package org.apache.catalina.startup;
 
 import com.techcourse.resolver.Resolver;
 import java.util.List;
 import com.techcourse.resolver.Dispatcher;
+import org.apache.catalina.startup.Server;
 
 public class WAS implements Server {
     private final Server tomcat;
