+++ tomcat/src/main/java/org/apache/coyote/http11/HttpCookie.java
@@ -1,4 +1,4 @@
-package com.techcourse;
+package org.apache.coyote.http11;
 
 import java.util.HashMap;
 import java.util.Map;
